scripts/requirements-timeline.txt
# Install into the repo venv so Cursor/Pylance match `python.defaultInterpreterPath`:
#   python3 -m venv .venv && .venv/bin/pip install -r scripts/requirements-timeline.txt
openai-whisper>=20231117
scenedetect[opencv]>=0.6.4
opencv-python>=4.8.0
